Refactor judge.js file, solves #24
